dep queries fix

+++ database/migrations/2023_11_13_134905_create_employee_table.php
@@ -37,7 +37,7 @@ return new class extends Migration
             $table->string('line_manager')->nullable();
             $table->string('contract_type')->nullable();
             $table->date('contract_renewal_date')->nullable();
-            $table->decimal('salary', 15)->nullable();
+            $table->double('salary', 15)->nullable();
             $table->string('currency')->nullable();
             $table->decimal('rate', 10)->nullable();
             $table->string('postal_address')->nullable();
